requirements.txt
Flask==3.0.2
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-mail==0.9.1
flask-bcrypt==1.0.1
python-dotenv==1.0.1
cryptography==42.0.5
pytest==8.0.2
python-magic-bin==0.4.14
gunicorn==21.2.0
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
python-magic==0.4.27 